refactor: move functionality regarding neural network to nnutils package
