+++ scrape_and_score/__main__.py
@@ -1,7 +1,5 @@
 from config import configure_logging
-from db import (
-    connection
-)
+from db import connection
 from db.insert.teams import insert_teams
 from db.read.teams import fetch_all_teams
 from config import load_configs, get_config
